Move 1.9 specific spec into a different file
